backend/requirements.txt
Django>=4.2.29,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Pillow>=12.1.1,<13.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
drf-spectacular>=0.27,<1.0
django-filter>=23.5,<24.0
PyJWT>=2.12.0,<3.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Linting
ruff>=0.2,<1.0
